+++ test/acceptance/tests/basic/terraform/basic-install/main.tf
@@ -210,8 +210,8 @@ EOT
   retry_join = [module.consul_server.server_dns]
   upstreams = [
     {
-      destination_name = "${var.server_service_name}_${var.suffix}"
-      local_bind_port  = 1234
+      destinationName = "${var.server_service_name}_${var.suffix}"
+      localBindPort   = 1234
     }
   ]
   log_configuration = local.test_client_log_configuration